Rename Shirmin.java to Duke.java and update runtest.bat accordingly
